+++ src/branch_predictor.c
@@ -20,17 +20,17 @@ u8 bp_predict(BRANCH_PREDICTOR* branch_predictor, ADDR pc, u8 result) {
     predicted = 0;
     #elif defined(BP_2BIT)
     u8 prev_stat = branch_predictor->counter;
-    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((signed)prev_stat - 1, 0);
+    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((s8)prev_stat - 1, 0);
     predicted = (prev_stat < 2) ? 0 : 1;
     branch_predictor->counter = curr_stat;
     #elif defined(BP_BIMODAL)
     u8 prev_stat = branch_predictor->pht[pc % PHT_SIZE];
-    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((signed)prev_stat - 1, 0);
+    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((s8)prev_stat - 1, 0);
     predicted = (prev_stat < 2) ? 0 : 1;
     branch_predictor->pht[pc % PHT_SIZE] = curr_stat;
     #elif defined(BP_GSHARE)
     u8 prev_stat = branch_predictor->pht[(pc ^ branch_predictor->gh) % PHT_SIZE];
-    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((signed)prev_stat - 1, 0);
+    u8 curr_stat = result ? min(prev_stat + 1, 3) : max((s8)prev_stat - 1, 0);
     predicted = (prev_stat < 2) ? 0 : 1;
     branch_predictor->pht[(pc ^ branch_predictor->gh) % PHT_SIZE] = curr_stat;
     branch_predictor->gh = (branch_predictor->gh << 1) | (result ? 1 : 0);
